+++ src/python/cpp/core.hpp
@@ -119,9 +119,9 @@ void define_core_bindings(py::module &m) {
         .def(py::init<>())
         .def_static("identity", &Transforms::identity)
         .def_static("zero", &Transforms::zero)
-        .def_static("reorder", py::overload_cast<NDView<size_t, 2> &>(&Transforms::reorder))
-        .def_static("reorder", py::overload_cast<NDArray<size_t, 2> &>(&Transforms::reorder))
-        .def_static("reorder", py::overload_cast<std::vector<size_t> &>(&Transforms::reorder))
+        .def_static("reorder", py::overload_cast<NDView<uint64_t, 2> &>(&Transforms::reorder))
+        .def_static("reorder", py::overload_cast<NDArray<uint64_t, 2> &>(&Transforms::reorder))
+        .def_static("reorder", py::overload_cast<std::vector<uint64_t> &>(&Transforms::reorder))
         .def_static(
             // only accept order_map of type uint64_t
             "reorder",
